Update avant_guard.ash
Burly Bodyguard should set the familiar as a 100% run. The 2nd run I did and I set it to a 100% Burly Bodyguard, it ran with minimal issues. But when I didn't set it on the 3rd run, it had so many problems that I had to handle a bunch of tasks manually. Autoscend kept trying to switch to another familiar.